Replace the sil-opt binary with a symlink to swift-frontend
rdar://76551283
