+++ lib/DriverTool/CMakeLists.txt
@@ -1,5 +1,6 @@
 set(driver_sources_and_options
                 driver.cpp
+                sil_opt_main.cpp
                 autolink_extract_main.cpp
                 modulewrap_main.cpp
                 swift_api_digester_main.cpp
